✨ feat(app): Add function execution when leave button is clicked on
Added an arbitrary void func() argument to Button::Update() so that action is performed (closing window in this case) when the button is clicked. Did this way to avoid hard coding the function inside the Button class and keep the code more modular. The downside of this approach is that rn buttons can only perform void, argument-free functions when clicked.